traffic-classification/requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
ultralytics>=8.3.0
